refactor: card컨테이너 분리
